Fixes for Jetson Nano board support.

+++ ggml/src/ggml-cuda/convert.cu
@@ -732,9 +732,9 @@ to_fp16_nc_cuda_t ggml_get_to_fp16_nc_cuda(ggml_type type) {
 to_bf16_nc_cuda_t ggml_get_to_bf16_nc_cuda(ggml_type type) {
     switch (type) {
         case GGML_TYPE_F32:
-            return convert_unary_cuda<float, nv_bfloat16>;
+            return convert_unary_cuda<float, half>;
         case GGML_TYPE_F16:
-            return convert_unary_cuda<half, nv_bfloat16>;
+            return convert_unary_cuda<half, half>;
         default:
             return nullptr;
     }
@@ -745,7 +745,7 @@ to_fp32_nc_cuda_t ggml_get_to_fp32_nc_cuda(ggml_type type) {
         case GGML_TYPE_F16:
             return convert_unary_cuda<half, float>;
         case GGML_TYPE_BF16:
-            return convert_unary_cuda<nv_bfloat16, float>;
+            return convert_unary_cuda<half, float>;
         default:
             return nullptr;
     }
